Revert "Refactor: rename `CreateEditDomainDialong`"
This reverts commit 082a1208b841820aadbd6561c0ddb3441c3471cf.
